refactor(extension): CHECKOUT-7652 Add namespace to extension command messages, remove duplicate commands and improve typing/naming
