+++ packages/core/src/extension/extension-commands.ts
@@ -4,31 +4,24 @@ export type ExtensionCommand =
     | SetIframeStyleCommand
     | FrameLoadedCommand;
 
-export const enum ExtensionCommandType {
-    FrameLoaded = 'FRAME_LOADED',
-    ReloadCheckout = 'RELOAD_CHECKOUT',
-    ShowLoadingIndicator = 'SHOW_LOADING_INDICATOR',
-    SetIframeStyle = 'SET_IFRAME_STYLE',
+export enum ExtensionCommandType {
+    FrameLoaded = 'FRAME_LOADED', // TODO: Define this as internal command/event
+    ReloadCheckout = 'EXTENSION:RELOAD_CHECKOUT',
+    ShowLoadingIndicator = 'EXTENSION:SHOW_LOADING_INDICATOR',
+    SetIframeStyle = 'EXTENSION:SET_IFRAME_STYLE',
 }
 
 export interface FrameLoadedCommand {
     type: ExtensionCommandType.FrameLoaded;
-    payload: {
-        extensionId: string;
-    };
 }
 
 export interface ReloadCheckoutCommand {
     type: ExtensionCommandType.ReloadCheckout;
-    payload: {
-        extensionId: string;
-    };
 }
 
 export interface ShowLoadingIndicatorCommand {
     type: ExtensionCommandType.ShowLoadingIndicator;
     payload: {
-        extensionId: string;
         show: boolean;
     };
 }
@@ -36,7 +29,6 @@ export interface ShowLoadingIndicatorCommand {
 export interface SetIframeStyleCommand {
     type: ExtensionCommandType.SetIframeStyle;
     payload: {
-        extensionId: string;
         style: {
             [key: string]: string | number | null;
         };
