+++ packages/core/src/extension/initialize-extension-service.ts
@@ -4,14 +4,15 @@ import {
     setupContentWindowForIframeResizer,
 } from '../common/iframe';
 
-import {
-    ExtensionCommand,
-    ExtensionCommandType,
-    ExtensionEventMap,
-    InitializeExtensionServiceOptions,
-} from './extension-client';
+import { ExtensionCommand, ExtensionCommandType } from './extension-commands';
+import { ExtensionEventMap } from './extension-events';
 import { ExtensionInternalCommand } from './extension-internal-commands';
 import ExtensionService from './extension-service';
+
+export interface InitializeExtensionServiceOptions {
+    extensionId: string;
+    parentOrigin: string;
+}
 
 export default function initializeExtensionService(options: InitializeExtensionServiceOptions) {
     const { extensionId, parentOrigin } = options;
@@ -25,7 +26,7 @@ export default function initializeExtensionService(options: InitializeExtensionS
     );
 
     extension.initialize(extensionId);
-    extension.post({ type: ExtensionCommandType.FRAME_LOADED, payload: { extensionId } });
+    extension.post({ type: ExtensionCommandType.FrameLoaded });
 
     return extension;
 }
