+++ packages/core/src/extension/subscribers/subscribe-consignments-change.ts
@@ -1,7 +1,7 @@
 import { CheckoutSelectors } from '../../checkout';
 import { DataStoreProjection } from '../../common/data-store';
-import { ExtensionEventType } from '../extension-client';
 import { ExtensionEventBroadcaster } from '../extension-event-broadcaster';
+import { ExtensionEventType } from '../extension-events';
 
 import { ExtensionChangeSubscriber } from './extension-change-subscriber';
